factor out CreateAssessmentTestsMixin into own file
