+++ mhep/mhep/dev/tests/views/mixins/__init__.py
@@ -1,1 +1,2 @@
 from .assert_error import AssertErrorMixin
+from .create_assessment_tests_mixin import CreateAssessmentTestsMixin
